additional auth handler callback to abstract verification of accounts
