+++ lib/Auth.js
@@ -89,23 +89,26 @@ function handleAuth(req, res, onSuccess, onFail, account, extraSalt) {
         return onFail();
       }
       else {
-        var posthash = crypto.createHash('sha1').update(user_info[jsh.map.user_hash].toString('hex') + account.tstmp + extraSalt).digest('hex');
-        if(jsh.Config.debug_params.auth_debug){
-          jsh.Log('Auth DB Hash:     '+posthash, { source: 'authentication' });
-          jsh.Log('Auth Client Hash: '+account.password, { source: 'authentication' });
-        }
-        if (posthash == account.password) {
-          req.isAuthenticated = true;
-          req.user_id = user_info[jsh.map.user_id];
-          req.user_name = req.jshsite.auth.getuser_name(user_info, jsh);
-          req._DBContext = req.jshsite.auth.getContextUser(user_info, jsh);
-          req._roles = {};
-          _.each(user_roles, function (role) { req._roles[role] = role; });
-          if(jsh.Config.debug_params.auth_debug) jsh.Log('Auth: Success', { source: 'authentication' });
-          if ('onAuthComplete' in req.jshsite.auth) req.jshsite.auth.onAuthComplete(req, user_info, jsh, onSuccess);
-          return onSuccess();
-        }
-        else return onFail();
+        req.jshsite.auth.on_validate(req, jsh, user_info, function(error, token){
+          if (error) return onFail();
+          var posthash = crypto.createHash('sha1').update(token + account.tstmp + extraSalt).digest('hex');
+          if(jsh.Config.debug_params.auth_debug){
+            jsh.Log('Auth DB Hash:     '+posthash, { source: 'authentication' });
+            jsh.Log('Auth Client Hash: '+account.password, { source: 'authentication' });
+          }
+          if (posthash == account.password) {
+            req.isAuthenticated = true;
+            req.user_id = user_info[jsh.map.user_id];
+            req.user_name = req.jshsite.auth.getuser_name(user_info, jsh);
+            req._DBContext = req.jshsite.auth.getContextUser(user_info, jsh);
+            req._roles = {};
+            _.each(user_roles, function (role) { req._roles[role] = role; });
+            if(jsh.Config.debug_params.auth_debug) jsh.Log('Auth: Success', { source: 'authentication' });
+            if ('onAuthComplete' in req.jshsite.auth) req.jshsite.auth.onAuthComplete(req, user_info, jsh, onSuccess);
+            return onSuccess();
+          }
+          else return onFail();
+        });
       }
     }
     else{
